perf(substreams): gate dynamic fee mapping by deployment block
