+++ protocols/substreams/base-aerodrome-slipstreams/src/modules/1_store_dynamic_fee_config.rs
@@ -1,20 +1,12 @@
 use crate::modules::utils::{
-    dynamic_fee_config_initialized_key, dynamic_fee_config_key, DynamicFeeEvent, Params,
+    dynamic_fee_config_initialized_key, dynamic_fee_config_key, should_process_dynamic_fee_config,
+    DynamicFeeEvent, Params,
 };
 use substreams::{
     scalar::BigInt,
     store::{StoreNew, StoreSet, StoreSetBigInt},
 };
 use substreams_ethereum::pb::eth::v2 as eth;
-
-// Earliest deployment among the configured fee modules:
-// - 0x090b2a6bb475c00e2256e2095a60887cd710803b at block 44_221_569
-// - 0xf4ecd78ebeb6d36cf7f80b5b6b41453515fe2785 at block 44_221_840
-const FIRST_DYNAMIC_FEE_MODULE_DEPLOYMENT_BLOCK: u64 = 44_221_569;
-
-fn should_process_dynamic_fee_config(block_number: u64) -> bool {
-    block_number >= FIRST_DYNAMIC_FEE_MODULE_DEPLOYMENT_BLOCK
-}
 
 fn set_config_value(
     store: &StoreSetBigInt,
@@ -51,21 +43,13 @@ pub fn store_dynamic_fee_config(params: String, block: eth::Block, store: StoreS
                 continue;
             };
             let pool = event.pool();
+            // Intentionally write the marker for every event. Only its first write has an empty
+            // old_value, which lets map_protocol_changes identify the pool's first dynamic-fee
+            // event and emit a complete snapshot.
             store.set(log.ordinal, dynamic_fee_config_initialized_key(pool), &BigInt::from(1));
             for (attribute, value) in event.config_updates() {
                 set_config_value(&store, log.ordinal, pool, attribute, &value);
             }
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::should_process_dynamic_fee_config;
-
-    #[test]
-    fn starts_processing_at_the_first_configured_fee_module_deployment() {
-        assert!(!should_process_dynamic_fee_config(44_221_568));
-        assert!(should_process_dynamic_fee_config(44_221_569));
     }
 }
